app/cars/requirements.txt
flask
connexion
requests 
grpc-interceptor ~= 0.11.0
grpcio-tools ~= 1.30
swagger-ui-bundle
SQLAlchemy==1.3.23
Flask-SQLAlchemy==2.1
connexion[swagger-ui]
grpc_pytools